Revert CompilationStatus
