--- GameServerCore/Enums/CompilationStatus.cs
@@ -1,11 +1,0 @@
-﻿namespace GameServerCore.Enums
-{
-    public enum CompilationStatus
-    {
-        Default = 0,
-        Compiled = 1,
-        SomeCompiled = 2,
-        NoneCompiled = 3,
-        NoScripts = 4
-    }
-}
